feat: Complete elimination of VATSIM dataclass dependencies
- Remove all VATSIMFlight, VATSIMController, VATSIMTransceiver dataclasses
- Convert data processing to pure dictionary-based approach
- Eliminate all attribute access (.flights, .controllers, .transceivers)
- Remove all hasattr() checks for dataclass attributes
- Simplify data flow: VATSIM API  Dictionary  Database
- Fix flight processing errors ('dict' object has no attribute errors)
- Maintain full functionality while improving performance
- Successfully tested with live VATSIM data ingestion

+++ database_audit/database_audit_framework.py
@@ -272,7 +272,7 @@ class DatabaseAuditor:
             'transceivers': [
                 ('callsign', 'VARCHAR(50)', 'Entity callsign'),
                 ('transceiver_id', 'INTEGER', 'Transceiver ID'),
-                ('frequency', 'INTEGER', 'Frequency in Hz'),
+                ('frequency', 'BIGINT', 'Frequency in Hz'),
                 ('position_lat', 'DOUBLE PRECISION', 'Latitude'),
                 ('position_lon', 'DOUBLE PRECISION', 'Longitude'),
                 ('height_msl', 'DOUBLE PRECISION', 'Height above sea level'),
